+++ metrohash64.cpp
@@ -25,7 +25,7 @@
 
 #include "metrohash.h"
 
-void metrohash64_1(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out)
+void metrohash64_1(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out)
 {
     static const uint64_t k0 = 0xC83A91E1;
     static const uint64_t k1 = 0x8648DBDB;
@@ -103,7 +103,7 @@ void metrohash64_1(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * o
 }
 
 
-void metrohash64_2(const uint8_t * key, uint64_t len, uint32_t seed, uint8_t * out)
+void metrohash64_2(const uint8_t * key, uint64_t len, uint64_t seed, uint8_t * out)
 {
     static const uint64_t k0 = 0xD6D018F5;
     static const uint64_t k1 = 0xA2AA033B;
